+++ rio_viz/io/__init__.py
@@ -1,3 +1,4 @@
 """rio_viz.io."""
 
+from rio_viz.io.mosaic import MosaicReader  # noqa
 from rio_viz.io.reader import MultiFilesReader  # noqa